Update based on PR comments

+++ packages/flutter/lib/src/material/mergeable_material.dart
@@ -138,7 +138,7 @@ class MergeableMaterial extends StatefulWidget {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties.add(EnumProperty<Axis>('mainAxis', mainAxis));
-    properties.add(DoubleProperty('elevation', elevation.toDouble()));
+    properties.add(DoubleProperty('elevation', elevation));
   }
 
   @override
@@ -615,7 +615,7 @@ class _MergeableMaterialState extends State<MergeableMaterial> with TickerProvid
 
     return _MergeableMaterialListBody(
       mainAxis: widget.mainAxis,
-      elevation: widget.elevation.toDouble(),
+      elevation: widget.elevation,
       items: _children,
       children: widgets,
     );
